+++ CHANGELOG.md
@@ -29,6 +29,7 @@
   - [x] fix GitHub Actions after upgrade to Spring Boot 3 (JDK17)
   - [x] fix `radp-spring-framework` ResponseBuilder bug
   - [x] fix META-INF/internal resource SPI
+  - [x] fix ApplicationContextHelper#getBean bug
 
 ## 3.12